+++ vaadincreate-ui/src/main/java/org/vaadin/tatu/vaadincreate/admin/UserManagementPresenter.java
@@ -5,6 +5,7 @@ import java.io.Serializable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.vaadin.tatu.vaadincreate.VaadinCreateUI;
+import org.vaadin.tatu.vaadincreate.auth.AccessControl;
 import org.vaadin.tatu.vaadincreate.backend.UserService;
 import org.vaadin.tatu.vaadincreate.backend.data.User;
 
@@ -13,6 +14,8 @@ public class UserManagementPresenter implements Serializable {
 
     private UserManagementView view;
     private UserService service = VaadinCreateUI.get().getUserService();
+    private AccessControl accessControl = VaadinCreateUI.get()
+            .getAccessControl();
 
     public UserManagementPresenter(UserManagementView view) {
         this.view = view;
@@ -24,6 +27,7 @@ public class UserManagementPresenter implements Serializable {
     }
 
     public void removeUser(int id) {
+        accessControl.assertAdmin();
         service.removeUser(id);
         logger.info("User '{}' removed.", id);
         view.showUserRemoved();
@@ -31,6 +35,7 @@ public class UserManagementPresenter implements Serializable {
     }
 
     public void updateUser(User user) {
+        accessControl.assertAdmin();
         try {
             service.updateUser(user);
             view.showUserUpdated();
